changing data types

+++ MyPlayground.playground/Contents.swift
@@ -1,4 +1,4 @@
 //Range operator
-let randomNum = Int.random(in:1..<100)
+let randomNum = Float.random(in:1..<100)
 //can use <= , >= etc.
 print(randomNum)
